requirements.txt
# Core framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.8.2               # 2.7.x requires pydantic-core source build (PyO3<3.13)

# LangChain / LangGraph
langchain==0.3.28             # 0.3.0 pins numpy<2.0 — incompatible with Python 3.13
langchain-openai==0.2.14
langgraph==0.2.76
langsmith==0.3.45

# Embeddings & reranking
FlagEmbedding==1.3.5         # BGE-M3 dense + sparse in one model; 1.3.2 pins transformers==4.44.2
sentence-transformers==3.0.0  # CrossEncoder reranker
torch==2.10.0

# AI detection (token entropy)
transformers==4.46.0          # 4.44.x requires tokenizers 0.19.x which has no Py3.13 wheel

# Vector store
qdrant-client==1.12.1             # 1.11.0 uses np.NINF which was removed in NumPy 2.0

# PII masking
presidio-analyzer==2.2.362
presidio-anonymizer==2.2.362
spacy==3.8.13                 # 3.7.x depends on blis which fails to compile on Python 3.13
# python -m spacy download en_core_web_lg

# Azure
azure-storage-blob==12.22.0
azure-servicebus==7.12.0
azure-keyvault-secrets==4.8.0
azure-identity==1.17.0

# Document parsing
pdfplumber==0.11.0
python-docx==1.1.0

# Caching / rate limiting
redis[asyncio]==5.0.0

# Monitoring
prometheus-client==0.16.0

# Utilities
numpy==2.1.0                  # 1.26.x has no Python 3.13 wheel
scikit-learn==1.7.0           # 1.5.x / 1.6.x have no Python 3.13 wheel
joblib==1.4.0                 # Model serialisation
python-multipart==0.0.20
